Add files via upload
This project is regarding social benefits especially women safety which I feel that it is the need of the hour.
Most of the sexual harassment are occurring at the city outskirts or unpopulated areas. If any women trapped in this situation she looks for help from others by calling 'help, help'. This is the case where mobile acts as intelligent device, hears the sound if "help" word detected it immediately records the sound of the environment and tracks the location and sends this information of sound ,location to police and their relatives , friends. So that we can avoid dangerous situations like sexual harassment, murder, women trafficking to some extent